requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.41
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.1.8
itsdangerous==2.2.0
blinker==1.9.0
greenlet==3.0.3 